# main/src/main/resources/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Policy Framework Lifecycle API
  description: The Policy Fraemwork API allows the lifecycle of policy types and policyes to be managed
  contact:
    name: ONAP Support
    url: https://lists.onap.org/g/onap-discuss
    email: onap-discuss@lists.onap.org
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: '1.0'
...
servers:
- url: //localhost:30440/
...
paths:
...
  /policies:
    get:
      tags:
      - "Policy Design"
      summary: Retrieve all versions of available policies
      description: Returns all version of available policies
      operationId: getPolicies
      parameters:
      - name: mode
        in: query
        description: Fetch mode for policies, BARE for bare  policies (default), REFERENCED
          for fully referenced policies
        schema:
          $ref: '#/components/schemas/PolicyFetchMode'
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: OK, returns the found policies in an instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
        401:
          description: Authentication Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        403:
          description: Authorization Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        404:
          description: Resource Not Found
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        500:
          description: Internal Server Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: Guilin
    post:
      tags:
      - "Policy Design"
      summary: Create one or more new policies
      description: Client should provide TOSCA body of the new polic(ies)
      operationId: createPolicies
      parameters:
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      requestBody:
        description: Entity body of the policies to be created contained in an instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToscaServiceTemplate'
          application/yaml:
            schema:
              $ref: '#/components/schemas/ToscaServiceTemplate'
        required: true
      responses:
        200:
          description: OK, the policies created in an instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
        201:
          description: OK, the policies created in an instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
        400:
          description: Invalid Body
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        401:
          description: Authentication Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        403:
          description: Authorization Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        404:
          description: Resource Not Found
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        406:
          description: Not Acceptable Payload
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        500:
          description: Internal Server Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: El Alto
      x-codegen-request-body-name: body
  /policies/{policyId}/versions/{policyVersion}:
    get:
      tags:
      - "Policy Design"
      summary: Retrieve specific version of a specified policy
      description: Returns a particular version of specified policy
      operationId: getSpecificPolicy
      parameters:
      - name: mode
        in: query
        description: Fetch mode for policies, BARE for bare  policies (default), REFERENCED
          for fully referenced policies
        schema:
          $ref: '#/components/schemas/PolicyFetchMode'
      - name: policyId
        in: path
        description: Name of policy
        required: true
        schema:
          type: string
      - name: policyVersion
        in: path
        description: Version of policy
        required: true
        schema:
          type: string
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: OK, returns the found policies in an instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
        401:
          description: Authentication Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        403:
          description: Authorization Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        404:
          description: Resource Not Found
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        500:
          description: Internal Server Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: Guilin
    delete:
      tags:
      - "Policy Design"
      summary: Delete a particular version of a policy
      description: 'Rule: the version that has been deployed in PDP group(s) cannot
        be deleted'
      operationId: deleteSpecificPolicy
      parameters:
      - name: policyId
        in: path
        description: ID of policy
        required: true
        schema:
          type: string
      - name: policyVersion
        in: path
        description: Version of policy
        required: true
        schema:
          type: string
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: OK, returns the deleted policy in an instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
        204:
          description: No Content
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content: {}
        401:
          description: Authentication Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        403:
          description: Authorization Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        404:
          description: Resource Not Found
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        409:
          description: Delete Conflict, Rule Violation
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        500:
          description: Internal Server Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: Guilin
...
  /policytypes/{policyTypeId}:
    get:
      tags:
      - "Policy Design"
      summary: Retrieve all available versions of a policy type
      description: Returns a list of all available versions for the specified policy
        type
      operationId: getAllVersionsOfPolicyType
      parameters:
      - name: policyTypeId
        in: path
        description: ID of policy type
        required: true
        schema:
          type: string
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: OK, the found policy types in an instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
        401:
          description: Authentication Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        403:
          description: Authorization Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        404:
          description: Resource Not Found
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        500:
          description: Internal Server Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: Dublin
  /policytypes/{policyTypeId}/versions/latest:
    get:
      tags:
      - "Policy Design"
      summary: Retrieve latest version of a policy type
      description: Returns latest version for the specified policy type
      operationId: getLatestVersionOfPolicyType
      parameters:
      - name: policyTypeId
        in: path
        description: ID of policy type
        required: true
        schema:
          type: string
      - name: X-ONAP-RequestID
        in: header
        description: RequestID for http transaction
        schema:
          type: string
          format: uuid
      responses:
        200:
          description: OK, the found policy types in an instance of
            [ToscaServiceTemplate](https://github.com/onap/policy-models/blob/master/models-tosca/src/main/java/org/onap/policy/models/tosca/authorative/concepts/ToscaServiceTemplate.java)
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-onap-RequestId:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
            application/yaml:
              schema:
                $ref: '#/components/schemas/ToscaServiceTemplate'
        401:
          description: Authentication Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        403:
          description: Authorization Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        404:
          description: Resource Not Found
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
        500:
          description: Internal Server Error
          headers:
            X-LatestVersion:
              schema:
                type: string
            X-PatchVersion:
              schema:
                type: string
            X-MinorVersion:
              schema:
                type: string
            X-ONAP-RequestID:
              schema:
                type: string
                format: uuid
          content: {}
      security:
      - basicAuth: []
      x-interface info:
        api-version: 1.0.0
        last-mod-release: Dublin
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    ToscaServiceTemplate:
      title: ToscaServiceTemplate
      type: object
...
    PolicyFetchMode:
      title: PolicyFetchMode
      type: object